Add helper method create and decrypt Stream.
